requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cssselect2==0.8.0
exceptiongroup==1.3.0
fastapi==0.115.12
fonttools==4.58.0
h11==0.16.0
httptools==0.6.4
idna==3.10
lxml==5.4.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
fastapi
uvicorn
weasyprint
python-docx
python-multipart
pydantic


